Implement function that can merge two list
Move element of list two to list one.
Compare the current value of list one and list two.
Current value of list two is always the first value.
Move the element of list two to list one if ascend/descend needed.
Or current element of list one becomes next element.
If current of list one is the last one or list two is empty.
Splice list two to the tail of list one.

Change-Id: Ie784ae445c06624e586c0e8af0d2decbc7be2425