Fix the range to DROP in the test17
Change insert head and tail and remove head and tail simulation 1.
In the original code, it only use N_MEASURES - DROP_SIZE * 2 input_data.
But it is prepared N_MEASURES input_data for it.
After differentiate all the input_data, I sort it with qsort in C.
And only Update Statics that is in the first (N_MEASURES-DROP_SIZE*2).
According to the paper "Dude, is my code constant time?".
"The upper tail may be more influenced by data-independent noise."
So I drop two DROP_SIZE of exec_time int the upper tail.

Change-Id: I4395d502e00cf16a199a1a01ff6266bf877eb155